Move unit tests into subfolder tests/unit
