+++ tests/unit/lib/AppTest.php
@@ -23,7 +23,7 @@ namespace OCA\Notifications\Tests\Lib;
 
 
 use OCA\Notifications\App;
-use OCA\Notifications\Tests\TestCase;
+use OCA\Notifications\Tests\Unit\TestCase;
 
 class AppTest extends TestCase {
 	/** @var \OCA\Notifications\Handler|\PHPUnit_Framework_MockObject_MockObject */
