+++ tests/unit/lib/HandlerTest.php
@@ -23,7 +23,7 @@ namespace OCA\Notifications\Tests\Lib;
 
 
 use OCA\Notifications\Handler;
-use OCA\Notifications\Tests\TestCase;
+use OCA\Notifications\Tests\Unit\TestCase;
 
 class HandlerTest extends TestCase {
 	/** @var \OCA\Notifications\Handler */
